requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
